# pathosync-api/swagger.yml
openapi: 3.0.3
info:
  title: Healthcare SaaS - Laboratory Test Management System API
  description: |
    API for a multi-tenant Laboratory Test Management System (India-focused).
    Use X-Tenant-ID header for tenant-specific operations and Bearer JWT for auth.
  version: "1.0.0"
servers:
  - url: https://your-api-domain.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PaginationPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PaginationLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
      description: Page size
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        country:
          type: string
    EmergencyContact:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
    MedicalHistoryItem:
      type: object
      properties:
        condition:
          type: string
        diagnosedDate:
          type: string
          format: date
        status:
          type: string
    Insurance:
      type: object
      properties:
        provider:
          type: string
        policyNumber:
          type: string
        validUntil:
          type: string
          format: date
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        alternatePhone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date
        age:
          type: integer
        gender:
          type: string
        bloodGroup:
          type: string
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContact'
        medicalHistory:
          type: array
          items:
            $ref: '#/components/schemas/MedicalHistoryItem'
        insurance:
          $ref: '#/components/schemas/Insurance'
        lastVisit:
          type: string
          format: date-time
        totalBills:
          type: integer
        totalAmount:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PatientListResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: '#/components/schemas/Patient'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreatePatientRequest:
      type: object
      required:
        - name
        - phone
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        bloodGroup:
          type: string
        emergencyContact:
          $ref: '#/components/schemas/EmergencyContact'
...
security:
  - bearerAuth: []
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: Get all patients with advanced filtering
      parameters:
        - $ref: '#/components/parameters/PaginationPage'
        - $ref: '#/components/parameters/PaginationLimit'
        - name: search
          in: query
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
            enum: [male, female, other]
        - name: ageRange
          in: query
          schema:
            type: string
        - name: registrationDate
          in: query
          schema:
            type: string
        - name: lastVisit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Patients list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'

    post:
      tags: [Patients]
      summary: Create new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
  /patients/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Patients]
      summary: Get patient details with history
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
    put:
      tags: [Patients]
      summary: Update patient information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patient'
      responses:
        "200":
          description: Updated patient
    delete:
      tags: [Patients]
      summary: Deactivate patient
      responses:
        "204":
          description: Patient deactivated
...